requirements.txt
# Core MCP and AI dependencies
mcp==1.11.0
anthropic==0.25.0
openai==1.30.0

# Web framework and async support
fastapi==0.110.0
uvicorn==0.27.0
pydantic==2.8.0

# Data processing and ML
pandas==2.2.0
numpy==1.26.0
scikit-learn==1.4.0
joblib==1.3.2

# Document processing
PyPDF2==3.0.1
pdf2image==1.16.3
pytesseract==0.3.10
Pillow==10.2.0

# Database and caching
redis==5.0.0
sqlalchemy==2.0.25
alembic==1.13.1

# Security and validation
cryptography==42.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Monitoring and logging
sentry-sdk==1.40.0
prometheus-client==0.19.0
structlog==23.2.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.0
pytest-cov==4.1.0
httpx==0.27.0

# Development tools
black==24.1.0
flake8==7.0.0
mypy==1.8.0
pre-commit==3.6.0
